split discovery actions to coin specific files
